requirements.txt
annotated-types==0.7.0
art==6.4
attrs==24.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
certifi==2024.8.30
charset-normalizer==3.4.0
editor==1.6.6
h11==0.14.0
idna==3.10
inquirer==3.4.0
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.0
orjson==3.10.12
outcome==1.3.0.post0
pandas==2.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
readchar==4.2.1
requests==2.32.3
rich==13.9.4
runs==1.2.2
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
xmod==1.8.1
